requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-Bcrypt==0.7.1
Flask-RESTful==0.3.9
Flask-Migrate==3.1.0
PyMySQL==1.0.2
python-dotenv==0.19.0
requests==2.26.0
Werkzeug==2.0.1
email-validator==1.1.3
WTForms==2.3.3 